AddHeadersPolicy.Factory -> AddHeadersPolicyFactory
